Add types to the private FloatingUI modifier
